Songs.swift
Path to the resource files are now in Songs.swift
Created new struct and an array that holds the title and file path for
each song.
In SongListController, the table view is now populated by this new
array.
In ViewController, the songs are now played from the song.swift file